Move loading of SDK projects to a helper class
